+++ src/sql/parse.py
@@ -1,3 +1,4 @@
+import re
 import itertools
 import shlex
 from os.path import expandvars
@@ -105,3 +106,42 @@ def without_sql_comment(parser, line):
 def magic_args(magic_execute, line):
     line = without_sql_comment(parser=magic_execute.parser, line=line)
     return parse_argstring(magic_execute, line)
+
+
+def escape_string_literals_with_colon_prefix(query):
+    """
+    Given a query, replaces all occurrences of ':variable' with '\:variable' and
+    ":variable" with "\:variable" so that the query can be passed to sqlalchemy.text
+    without the literals being interpreted as bind parameters. It doesn't replace
+    the occurrences of :variable (without quotes)
+    """  # noqa
+
+    # Define the regular expression pattern for valid Python identifiers
+    identifier_pattern = r"\b[a-zA-Z_][a-zA-Z0-9_]*\b"
+
+    double_quoted_variable_pattern = r'(?<!\\)":(' + identifier_pattern + r')(?<!\\)"'
+
+    # Define the regular expression pattern for matching ':variable' format
+    single_quoted_variable_pattern = r"(?<!\\)':(" + identifier_pattern + r")(?<!\\)\'"
+
+    # Replace ":variable" and ':variable' with "\:variable"
+    query_quoted = re.sub(double_quoted_variable_pattern, r'"\\:\1"', query)
+    query_quoted = re.sub(single_quoted_variable_pattern, r"'\\:\1'", query_quoted)
+
+    double_found = re.findall(double_quoted_variable_pattern, query)
+    single_found = re.findall(single_quoted_variable_pattern, query)
+
+    return query_quoted, double_found + single_found
+
+
+def find_named_parameters(input_string):
+    # Define the regular expression pattern for valid Python identifiers
+    identifier_pattern = r"\b[a-zA-Z_][a-zA-Z0-9_]*\b"
+
+    # Define the regular expression pattern for matching :variable format
+    variable_pattern = r'(?<!["\'])\:(' + identifier_pattern + ")"
+
+    # Use findall to extract all matches of :variable from the input string
+    matches = re.findall(variable_pattern, input_string)
+
+    return matches