+++ src/sql/warnings.py
@@ -1,2 +1,6 @@
 class JupySQLDataFramePerformanceWarning(UserWarning):
     pass
+
+
+class JupySQLQuotedNamedParametersWarning(UserWarning):
+    pass